parkrun/requirements.txt
Flask==3.1.0
google-cloud-firestore==2.19.0
google-cloud-secret-manager==2.20.2
gunicorn==21.2.0